use css modules, get rid of store-resources dep, get rid of R.prop, move Rebuy to index
