+++ react/rebuy.css
@@ -1,0 +1,11 @@
+.item {}
+.items {}
+.itemColumn {}
+.itemDetails {}
+.itemName {}
+.itemPrice {}
+
+
+.totalContainer {}
+.totalTitle {}
+.totalPrice {}
